googleSamples#26 Change default url http to https
The default 'http' doesn't show the icon feature of the shortcut, due to URLConnection not following cross-protocol redirects, e.g. http -> https.

Changed the default from http to https, and updated the screen shot.

Note, the screen shot in this commit still doesn't display in README.md due to googleSamples#23.